feat: Add map view
