+++ package.json
@@ -12,6 +12,7 @@
         "test": "vitest run .",
         "testw": "vitest",
         "lint": "eslint .",
+        "lf": "eslint .",
         "preview": "vite preview --port 4173 --host",
         "pwa-test": "bun run build && bun run preview",
         "fc": "prettier . -c",
@@ -20,11 +21,10 @@
         "fwf": "prettier -w",
         "el": "eslint .",
         "elf": "eslint . --fix",
-        "format": "eslint . --fix && prettier . -w",
+        "fmt": "eslint . --fix && prettier . -w",
         "build": "tsc -b && vite build",
         "vc": "eslint . && prettier . -c && vitest run && tsc -b --verbose && vite build",
-        "deploy": "bun run format && bun run vc",
-        "dp": "bun run deploy",
+        "dp": "bun run fmt && bun run vc",
         "gen-mod": "bunx plop module-gen",
         "unused": "knip"
     },
