+++ src/modules/branch/components/forms/create-branch-form.tsx
@@ -10,7 +10,7 @@ import { IMedia } from '@/modules/media'
 import { CountryCombobox } from '@/components/comboboxes/country-combobox'
 import { GradientBackground } from '@/components/gradient-background/gradient-background'
 import { HouseIcon, LoadingSpinnerIcon } from '@/components/icons'
-import MapPicker from '@/components/map/map-picker/map-picker'
+import MapPicker from '@/components/map/map-picker'
 import Modal, { IModalProps } from '@/components/modals/modal'
 import TextEditor from '@/components/text-editor'
 import { Button } from '@/components/ui/button'
